refactor(repos): rename base_repository to InMemoryRepository classes
Rename BaseReadRepository -> InMemoryReadRepository,
BaseWriteRepository -> InMemoryWriteRepository,
BaseRepository -> InMemoryRepository to reveal in-memory HashMap
storage intent.

Also apply review-driven fixes:
- Add cooperative super().__init__() in both parent __init__ methods
  to ensure diamond MRO chain executes correctly
- Remove misleading __slots__ declarations (no ancestor slots)

Update __init__.py and aggregate_repository imports.
