+++ src/forging_blocks/infrastructure/repositories/__init__.py
@@ -1,15 +1,13 @@
 """In-memory repository implementations for the infrastructure layer."""
 
 from .aggregate_repository import AggregateRepository
-from .base_repository import BaseReadRepository, BaseRepository, BaseWriteRepository
 from .in_memory_read_repository import InMemoryReadRepository
+from .in_memory_repository import InMemoryRepository
 from .in_memory_write_repository import InMemoryWriteRepository
 
 __all__ = [
     "AggregateRepository",
-    "BaseReadRepository",
-    "BaseRepository",
-    "BaseWriteRepository",
     "InMemoryReadRepository",
+    "InMemoryRepository",
     "InMemoryWriteRepository",
 ]
